backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
Pillow>=10.0.0
numpy>=1.26.0,<2.0.0
opencv-python>=4.8.0
python-multipart>=0.0.9
